Add atop service script (/service/run)
Add service script that runs atop in daemon mode, logging to
/var/log/atop/atop-{YYmmdd}.

The atop service is restarted every day at midnight, which triggers
the creation of a new log file. This is achieved using a separate
`atop-rotate` service (which is uses `snooze`).

We use s6 as process supervisor, so that the atop daemon and the
rotate service can be run in a single Docker container (in Kubernetes,
we could use two containers in a pod, instead).